Update and rename README.md to README.md
The purpose of the project is to provide online facility to 
Institutes to conduct online exams and to Students to give online exams. Institutes can enter and edit the questions along with the students list. Also they can view the result. Students can login and give their respective exams and view their score then and there.Others can view sample papers to get look and feel of the online examination system. 
1.2 Scope: The website to conduct online Admission for a University is “Online Admission System”. This website provides facility to institutes to conduct online Admissions by providing a unique id to each Student. The institute provides online registration of the student. Institute also enters the list of eligible students. All the information entered can be later edited by the institute. In turn student can login with their id, name and Programmed to give the forms and can view their Admission status then and there. Institutes can also change the status of admission of their students (if needed).